feat: adapt Member to conform to enum too and change value holded into simply pointer holded
