+++ include/GIL/Instructions/Member.hpp
@@ -3,36 +3,29 @@
 
 #include "AST/Types.hpp"
 #include "Type.hpp"
+#include "Value.hpp"
 
 namespace glu::gil {
 
-class Value;
-
 /// @class Member
-/// @brief Represents a member (field) of a structure in the Glu language.
+/// @brief Represents a member (field) of a structure or an enum in the Glu
+/// language.
 ///
 /// This class encapsulates the necessary information to represent and access
 /// a struct member in the GLU GIL (Generic Intermediate Language).
 class Member {
-    Value *_parentStruct; ///< The value of the parent structure
-    Value *_value; ///< Pointer to the value contained in this member
+    Type _type; ///< Pointer to the type contained in this member
     llvm::StringRef _name; ///< The name of the member
 
 public:
     /// @brief Constructs a Member object.
     ///
-    /// @param parentStruct The type of the parent structure
     /// @param field The field metadata information
     /// @param value Pointer to the value contained in this member
-    Member(Value *parentStruct, Value *value, llvm::StringRef name)
-        : _parentStruct(parentStruct), _value(value), _name(name)
+    Member(Value parent, Type type, llvm::StringRef name)
+        : _type(type), _name(name)
     {
     }
-
-    /// @brief Gets the parent structure type.
-    ///
-    /// @return The type of the parent structure.
-    Value *getParentStruct() const { return _parentStruct; }
 
     /// @brief Gets the field metadata.
     ///
@@ -42,7 +35,7 @@ public:
     /// @brief Gets the value contained in this member.
     ///
     /// @return Pointer to the value contained in this member.
-    Value *getValue() const { return _value; }
+    Type getType() const { return _type; }
 };
 
 }
